Panel administratora - usuwanie użytkowników i nadawanie ról

+++ HybrydowaApp/src/main/java/com/example/demo/controllers/BookController.java
@@ -1,12 +1,15 @@
 package com.example.demo.controllers;
 
 import com.example.demo.models.Book;
+import com.example.demo.models.User;
 import com.example.demo.repository.BookRepository;
+import com.example.demo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -14,6 +17,9 @@ public class BookController {
 
     @Autowired
     BookRepository bookRepository;
+
+    @Autowired
+    UserRepository userRepository;
 
     @GetMapping("/books")
     public List<Book> getBooks() {
@@ -44,10 +50,22 @@ public class BookController {
     @GetMapping("/bookDelete")
     void deleteBook(@RequestParam long id){
         Optional<Book> book = bookRepository.findById(id);
+        AtomicBoolean isForeignKey = new AtomicBoolean(false);
 
         if(book.isPresent()){
-            bookRepository.delete(book.get());
-            System.out.println("Pomyślnie usunięto książkę");
+            //książka jest kluczem obcym w innej tabeli
+            userRepository.findAll().forEach(x -> {
+                if(x.getBooks().contains(book.get())){
+                    isForeignKey.set(true);
+                }
+            });
+
+            if(isForeignKey.get()){
+                System.out.println("Nie można usunąć książki. Jest ona przypisana do przynajmniej jednego uzytkownika");
+            } else {
+                bookRepository.delete(book.get());
+                System.out.println("Pomyślnie usunięto książkę");
+            }
         }
         else {
             System.out.println("Nie znaleziono książki poprzez id");
